+++ server/types/app.go
@@ -49,20 +49,10 @@ type (
 		RegisterTxService(client.Context)
 
 		// RegisterTendermintService registers the gRPC Query service for tendermint queries.
-		RegisterTendermintService(client.Context)
+		RegisterTendermintService(clientCtx client.Context)
 
-		// CommitMultiStore Returns the multistore instance
+		// Return the multistore instance
 		CommitMultiStore() sdk.CommitMultiStore
-	}
-
-	// ApplicationQueryService defines an extension of the Application interface
-	// that facilitates gRPC query Services.
-	//
-	// NOTE: This interfaces exists only in the v0.45.x line to ensure the existing
-	// Application interface does not introduce API breaking changes.
-	ApplicationQueryService interface {
-		// RegisterNodeService registers the node gRPC Query service.
-		RegisterNodeService(client.Context)
 	}
 
 	// AppCreator is a function that allows us to lazily initialize an